+++ src/main/java/com/project/lecture/api/Listen/controller/ListenController.java
@@ -1,9 +1,9 @@
-package com.project.lecture.Listen.controller;
+package com.project.lecture.api.Listen.controller;
 
 import static com.project.lecture.type.ResponseType.COURSE_SUCCESS;
 import static com.project.lecture.type.ResponseType.DELETE_SUCCESS;
 
-import com.project.lecture.Listen.application.ListenApplication;
+import com.project.lecture.api.Listen.application.ListenApplication;
 import java.security.Principal;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
