+++ src/main/java/com/project/lecture/api/Listen/service/ListenService.java
@@ -1,4 +1,4 @@
-package com.project.lecture.Listen.service;
+package com.project.lecture.api.Listen.service;
 
 import com.project.lecture.entity.Listening;
 import com.project.lecture.repository.ListeningRepository;
