+++ src/main/java/com/project/lecture/api/course/application/CourseApplication.java
@@ -1,14 +1,14 @@
-package com.project.lecture.course.application;
+package com.project.lecture.api.course.application;
 
 
-import com.project.lecture.course.dto.CourseDto;
-import com.project.lecture.course.dto.CourseRequest.Create;
-import com.project.lecture.course.dto.CreateLecture;
-import com.project.lecture.course.service.CourseService;
-import com.project.lecture.course.service.LectureService;
+import com.project.lecture.api.course.dto.CourseDto;
+import com.project.lecture.api.course.dto.CourseRequest.Create;
+import com.project.lecture.api.course.dto.CreateLecture;
+import com.project.lecture.api.course.service.CourseService;
+import com.project.lecture.api.course.service.LectureService;
 import com.project.lecture.entity.Course;
 import com.project.lecture.entity.Member;
-import com.project.lecture.user.service.MemberService;
+import com.project.lecture.api.user.service.MemberService;
 import java.util.List;
 import java.util.stream.Collectors;
 import lombok.RequiredArgsConstructor;
