+++ src/main/java/com/project/lecture/api/user/application/SignUpApplication.java
@@ -1,8 +1,8 @@
-package com.project.lecture.user.application;
+package com.project.lecture.api.user.application;
 
+import com.project.lecture.api.user.dto.UserRequest.SignUp;
 import com.project.lecture.exception.kind.ExceptionExistUser;
-import com.project.lecture.user.dto.UserRequest;
-import com.project.lecture.user.service.MemberService;
+import com.project.lecture.api.user.service.MemberService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
@@ -12,7 +12,7 @@ import org.springframework.stereotype.Service;
 public class SignUpApplication {
   private final PasswordEncoder passwordEncoder;
   private final MemberService memberService;
-  public void saveUserByReq(UserRequest.SignUp sign) {
+  public void saveUserByReq(SignUp sign) {
     if (memberService.hasEmail(sign.getEmail())) {
       throw new ExceptionExistUser(sign.getEmail());
     }
