+++ src/test/java/com/project/lecture/Listen/application/ListenApplicationTest.java
@@ -10,15 +10,16 @@ import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
 import com.project.lecture.Helper.CommonHelper;
-import com.project.lecture.Listen.service.ListenService;
-import com.project.lecture.course.service.CourseService;
+import com.project.lecture.api.Listen.application.ListenApplication;
+import com.project.lecture.api.Listen.service.ListenService;
+import com.project.lecture.api.course.service.CourseService;
 import com.project.lecture.entity.Course;
 import com.project.lecture.entity.Member;
 import com.project.lecture.exception.SuperException;
 import com.project.lecture.exception.kind.ExceptionExistCourse;
 import com.project.lecture.exception.kind.ExceptionNotFoundCourse;
-import com.project.lecture.planner.service.PlannerService;
-import com.project.lecture.user.service.MemberService;
+import com.project.lecture.api.planner.service.PlannerService;
+import com.project.lecture.api.user.service.MemberService;
 import org.junit.Assert;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
