+++ src/test/java/com/project/lecture/Listen/service/ListenServiceTest.java
@@ -9,6 +9,7 @@ import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
 import com.project.lecture.Helper.CommonHelper;
+import com.project.lecture.api.Listen.service.ListenService;
 import com.project.lecture.entity.Listening;
 import com.project.lecture.repository.ListeningRepository;
 import org.junit.jupiter.api.DisplayName;
