+++ src/test/java/com/project/lecture/course/application/CourseApplicationTest.java
@@ -9,13 +9,14 @@ import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
 import com.project.lecture.Helper.CommonHelper;
-import com.project.lecture.course.dto.CourseDto;
-import com.project.lecture.course.dto.CourseRequest;
-import com.project.lecture.course.service.CourseService;
-import com.project.lecture.course.service.LectureService;
+import com.project.lecture.api.course.application.CourseApplication;
+import com.project.lecture.api.course.dto.CourseDto;
+import com.project.lecture.api.course.dto.CourseRequest;
+import com.project.lecture.api.course.service.CourseService;
+import com.project.lecture.api.course.service.LectureService;
 import com.project.lecture.entity.Course;
 import com.project.lecture.entity.Member;
-import com.project.lecture.user.service.MemberService;
+import com.project.lecture.api.user.service.MemberService;
 import java.util.List;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
