+++ src/test/java/com/project/lecture/course/service/LectureServiceTest.java
@@ -6,6 +6,7 @@ import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
 import com.project.lecture.Helper.CommonHelper;
+import com.project.lecture.api.course.service.LectureService;
 import com.project.lecture.entity.Lecture;
 import com.project.lecture.repository.LectureRepository;
 import java.util.List;
