+++ src/test/java/com/project/lecture/planner/service/PlannerServiceTest.java
@@ -11,6 +11,7 @@ import static org.mockito.Mockito.times;
 import static org.mockito.Mockito.verify;
 
 import com.project.lecture.Helper.CommonHelper;
+import com.project.lecture.api.planner.service.PlannerService;
 import com.project.lecture.entity.Lecture;
 import com.project.lecture.entity.Planner;
 import com.project.lecture.repository.PlannerRepository;
