+++ src/test/java/com/project/lecture/user/service/MemberServiceTest.java
@@ -7,6 +7,7 @@ import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
 import com.project.lecture.Helper.CommonHelper;
+import com.project.lecture.api.user.service.MemberService;
 import com.project.lecture.entity.Member;
 import com.project.lecture.exception.SuperException;
 import com.project.lecture.repository.MemberRepository;
